--- src/Doturn.Common/Program.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace Doturn.StunMessage
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello World!");
-        }
-    }
-}
