--- src/Doturn.StunAttribute/Program.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace Doturn.StunAttribute
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello World!");
-        }
-    }
-}
