+++ src/Doturn/StunAttribute/Nonce.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Linq;
-using Doturn.Common;
 
 namespace Doturn.StunAttribute
 {
