+++ src/Doturn/StunAttribute/Type.cs
