+++ src/Doturn/StunMessage/StunMessageBase.cs
@@ -1,0 +1,7 @@
+namespace Doturn.StunMessage
+{
+    abstract class StunMessageBase
+    {
+
+    }
+}
