+++ test/Doturn.Test/Doturn.Test.csproj
@@ -19,4 +19,8 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\Doturn\Doturn.csproj" />
+  </ItemGroup>
+
 </Project>
