+++ test/Doturn.Test/StunAttribute/LifetimeTest.cs
