+++ test/Doturn.Test/StunAttribute/TypeTest.cs
