+++ test/Doturn.Test/StunMessage/TypeTest.cs
